# backend/tutorlink-backend/openapi.json
{"openapi":"3.0.1","info":{"title":"TutorLink API","description":"API pública para la integración con el frontend de TutorLink","version":"v1"},"servers":[{"url":"http://localhost:18081","description":"Generated server url"}],"paths":{"/usuarios/{id}/foto":{"put":{"tags":["user-controller"],"operationId":"actualizarFoto","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/admin/usuarios/{id}/estado":{"put":{"tags":["admin-controller"],"operationId":"cambiarEstado","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"activo","in":"query","required":true,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/respuestas/{id}/rechazar":{"post":{"tags":["answer-controller"],"operationId":"rechazarYRegenerar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AnswerDTO"}}}}}}},"/respuestas/{id}/aprobar":{"post":{"tags":["answer-controller"],"operationId":"aprobar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AnswerDTO"}}}}}}},"/preguntas":{"post":{"tags":["question-controller"],"operationId":"crearPregunta","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuestionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/QuestionDTO"}}}}}}},"/preguntas/{id}/generar":{"post":{"tags":["answer-controller"],"operationId":"generarRespuestaLLM","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idUsuarioLLM","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AnswerDTO"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/admin/usuarios/{id}/aprobar":{"post":{"tags":["admin-controller"],"operationId":"aprobarUsuario","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/admin/estudiantes/{idEst}/asignar-tutor/{idTutor}":{"post":{"tags":["admin-controller"],"operationId":"asignarTutor","parameters":[{"name":"idEst","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idTutor","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}},"/usuarios/{id}":{"get":{"tags":["user-controller"],"operationId":"obtenerPerfil","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/usuarios/tutores/{idTutor}/tutorados":{"get":{"tags":["user-controller"],"operationId":"obtenerTutorados","parameters":[{"name":"idTutor","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"/usuarios/tutores/{idTutor}/tutorados/filtrar":{"get":{"tags":["user-controller"],"operationId":"filtrarTutorados","parameters":[{"name":"idTutor","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idPlan","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"semestre","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"/usuarios/tutores/{idTutor}/tutorados/buscar":{"get":{"tags":["user-controller"],"operationId":"buscarTutorados","parameters":[{"name":"idTutor","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"q","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}}},"/tutores/{idTutor}/pendientes":{"get":{"tags":["question-controller"],"operationId":"pendientesTutor","parameters":[{"name":"idTutor","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionDTO"}}}}}}}},"/preguntas/sugerencias":{"get":{"tags":["question-controller"],"operationId":"sugerencias","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionDTO"}}}}}}}},"/preguntas/buscar":{"get":{"tags":["question-controller"],"operationId":"buscar","parameters":[{"name":"q","in":"query","required":false,"schema":{"type":"string"}},{"name":"scope","in":"query","required":false,"schema":{"type":"string"}},{"name":"idFacultad","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"idPlan","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionDTO"}}}}}}}},"/admin/preguntas/{id}":{"delete":{"tags":["admin-controller"],"operationId":"eliminarPregunta","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponse"}}}}}}}},"components":{"schemas":{"UpdateProfileRequest":{"required":["fotoPerfilUrl"],"type":"object","properties":{"fotoPerfilUrl":{"type":"string"}}},"UserDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"apellidos":{"type":"string"},"correo":{"type":"string"},"rol":{"type":"string"},"fotoPerfilUrl":{"type":"string"},"telefono":{"type":"string"}}},"ApiResponse":{"type":"object","properties":{"ok":{"type":"boolean"},"mensaje":{"type":"string"}}},"AnswerDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"preguntaId":{"type":"integer","format":"int64"},"contenido":{"type":"string"},"version":{"type":"integer","format":"int32"},"estado":{"type":"string"},"fecha":{"type":"string","format":"date-time"},"aprobadorId":{"type":"integer","format":"int64"}}},"CreateQuestionRequest":{"required":["scopeTipo","texto","titulo"],"type":"object","properties":{"titulo":{"type":"string"},"texto":{"type":"string"},"scopeTipo":{"type":"string"},"idFacultadScope":{"type":"integer","format":"int64"},"idPlanEducativoScope":{"type":"integer","format":"int64"}}},"QuestionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"texto":{"type":"string"},"estado":{"type":"string"},"fechaCreacion":{"type":"string","format":"date-time"},"scopeTipo":{"type":"string"},"facultadId":{"type":"integer","format":"int64"},"planEducativoId":{"type":"integer","format":"int64"},"autorId":{"type":"integer","format":"int64"},"autorNombre":{"type":"string"}}},"RegisterRequest":{"required":["apellidos","contrasena","correo","nombre","rol"],"type":"object","properties":{"correo":{"type":"string"},"contrasena":{"maxLength":2147483647,"minLength":6,"type":"string"},"nombre":{"type":"string"},"apellidos":{"type":"string"},"rol":{"type":"string"},"matricula":{"type":"string"},"idPlanEducativo":{"type":"integer","format":"int64"},"idTutorAsignado":{"type":"integer","format":"int64"},"telefono":{"type":"string"},"idProgramaEducativo":{"type":"integer","format":"int64"}}},"LoginRequest":{"required":["contrasena","correo"],"type":"object","properties":{"correo":{"type":"string"},"contrasena":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"usuario":{"$ref":"#/components/schemas/UserDTO"}}}}}}